backend/Project/requirements.txt
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==5.1.0
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
crontab==0.22.8
cryptography==38.0.4
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.4
django-braces==1.15.0
django-channels==0.7.0
django-cors-headers==3.13.0
django-filter==22.1
django-jinja==2.10.2
django-oauth-toolkit==2.2.0
django-webpush==0.3.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==1.2.1
drf-yasg==1.21.4
ecdsa==0.18.0
filelock==3.8.2
http-ece==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.4.2
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
platformdirs==2.6.0
py-vapid==1.9.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.6
pywebpush==1.9.4
PyYAML==6.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.1
wrapt==1.14.1